requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
packaging==23.2
pip-review==1.3.0
Pygments==2.17.2
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
